minor readme + install tweaks
I discovered that the codebase currently doesn't work for ruby < 2.2.
The main problem is that the CXXFLAGS variable was added to mkmf in 2.2,
which might be a fixable problem, but for now it's easier to just
require users to have >= 2.2.

This also adds a missing `bundle install` line to the install steps.