Fix a segfault in Session run
Errors were ignored in the C++ implementation of Session_Run, which
meant that the ruby code could segfault as it would try to continue on
after a failed run.

This fix ensures Session_Run will throw an error that can be caught in
ruby to avoid segfaulting for cases like passing in a tensor with an
improper shape.